module_2b/requirements.txt
pika==1.3.2
mysql-connector-python==9.0.0
psycopg2-binary==2.9.9